% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CIAdata}
\alias{CIAdata}
\title{Return a dataset based on the CIA World Factbook}
\usage{
CIAdata(code = NULL)
}
\arguments{
\item{code}{An optional parameter specifying the code of the desired dataset}
}
\description{
This function can be used in two different ways. Without an argument, it returns a reference
table that includes information about all the CIA World Factbook tables that are available
through this function. Note the  \code{Code} column that indicates a unique code for each
available dataset. If this code is passed as an argument to the function, the function
will return the corresponding dataset.
}

