\name{countryCentroids}
\alias{countryCentroids}
\docType{data}
\title{
Point locations of countries
}
\description{
Gives the latitude/longitude centroid for each country.  This is useful
for making dot or other geom_point layers for each country.
}
\usage{data("countryCentroids")}
\format{
  A data frame with 241 observations on the following 6 variables.
  \describe{
    \item{\code{name}}{common names of countries}
    \item{\code{iso_a3}}{country codes in ISO_a3 format: unique three letter code}
    \item{\code{long}}{longitude of the country's centroid}
    \item{\code{lat}}{latitude}
    \item{\code{group}}{needed for ggplot compatibility (really?)}
    \item{\code{order}}{needed for ggplot compatibility (really?)}
  }
}
\details{
This was generated by by averaging the polygon boundry points for each country.
}
\source{
The original shapefile comes from \code{www.naturalearthdata.com/download/50m/cultural/ne_50m_admin_0_countries.zip}
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
\dontrun{
population <- CIAdata(2119)
population2 <- left_join(population,mutate(countryCentroids,country=name))
p <- makeWorldMap(population,'country')
p +  geom_polygon( color='gray', fill=NA) + # for the boundaries
  geom_point(data=population2, aes(x=long,y=lat,size=sqrt(pop)))
}
}
\keyword{datasets}
