\name{DCFdevel-package}
\alias{DCFdevel-package}
\alias{DCFdevel}
\docType{package}
\title{
Testing organization of Materials for the DCF course
}
\description{
This is a package to let me try out different ideas for organizing the DCF course
}
\details{
\tabular{ll}{
Package: \tab DCFdevel\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-22\cr
License: \tab Do not use.  All rights reserved.\cr
}
}
\author{
Daniel Kaplan

Maintainer: Daniel Kaplan <dtkaplan@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
