% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getJoinData}
\alias{getJoinData}
\title{Return the complete dataset used to plot the corresponding map}
\usage{
getJoinData(data, varname.x)
}
\arguments{
\item{data}{A dataset with countries as cases}

\item{varname.x}{The variable name for the column that holds the countries names}
}
\description{
Return the complete dataset used to plot the corresponding map
}

