% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeWorldMap}
\alias{makeWorldMap}
\title{Create a ggplot2 frame with a world map.}
\usage{
makeWorldMap(data, varname.x)
}
\arguments{
\item{data}{A dataset with countries as cases}

\item{varname.x}{The variable name for the column that holds the countries names}
}
\value{
A ggplot map object with no layers.
}
\description{
This function takes in a dataset with countries as cases.
Provide the country name as a three-letter iso_a3 or any common name
for the country (e.g. Canada, Fiji).
}
\details{
It standardizes the
country names and merges it another dataset which includes geographic coordinates
for each country. It returns a ggplot map with no layers. It prints a message
indicating how many countries will not be plotted because no latitude /
longitude data was found.
}
\examples{
\dontrun{
somedat <- CIAdata(2001) # 2001 codes for GDP
p <- makeWorldMap(somedat,'country')
p + geom_polygon(col="blue',aes(fill=GDP))
}
}

