% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standCountry}
\alias{standCountry}
\title{Standardize country names in a dataset}
\usage{
standCountry(data, varname.x, nonMatch = c(FALSE, TRUE))
}
\arguments{
\item{data}{A dataset with countries as cases}

\item{varname.x}{The variable name for the column that holds the countries names}

\item{nonMatch}{A boolean specifying whether or not a list of the countries without
a match should be returned by the function}
}
\value{
The updated dataset with standardized names (\code{$Data}) and the list of
countries without a match (\code{$Unmatched}) if the argument \code{nonMatch} is
set to \code{TRUE}
}
\description{
This function is built-in the \code{MakeWorldMap} function, so you shouldn't
need to use directly.
}

