\name{WorldCities}
\alias{WorldCities}
\docType{data}
\title{Cities and their populations}
\description{
A list of cities
}
\usage{data(WorldCities)}
\format{
  A data frame with 23018 observations on the following 10 variables.
  \describe{
    \item{\code{Code}}{The ISO (?) city code}
    \item{\code{Name}}{Name of the city}
    \item{\code{Latitude}}{location in degrees}
    \item{\code{Longitude}}{location in degrees}
    \item{\code{Country}}{Two letter country code}
    \item{\code{CountryRegion}}{A numerical region}
    \item{\code{Population}}{Population}
    \item{\code{RegionCode}}{ISO (?) Code}
    \item{\code{Region}}{Name of the region}
    \item{\code{Date}}{Date estimate made}
  }
}
\source{
In Draft: Somewhere on the Internet.  We need a proper source.
}
\references{
In Draft: We need a proper reference
}
\examples{
data(WorldCities)
## maybe str(WorldCities) ; plot(WorldCities) ...
}
\keyword{datasets}
